fix: bibgrep fulltext search
The parameters given to `print_key()` are "filename:nbr:line" where
 * filename: name of bib file
 * nbr: line number of the beginning of the entry
 * full line of text (beginning of entry)
This was automatically created by `egrep -n`.

Now, the fulltext search must create a similar line to re-use the
`print_key()` function. So far, the third parameter was the bib key,
but this commit changes it to the full line (of the last/current bib key).